+++ src/common/interfaces/workflows/index.ts
@@ -1,2 +1,3 @@
 export * from './token.interface';
 export * from './state.interface';
+export * from './data-flow.interface';