+++ src/common/models/artifact.model.ts
@@ -2,5 +2,6 @@ import { BaseModel } from './base.model';
 
 export class ArtifactModel<Meta> extends BaseModel<ArtifactModel<Meta>, Meta> {
   identity!: string;
+  type!: string;
   metadata!: Meta;
 }