+++ CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
+- Workflow DataFlow type added
+### Changed
+- Notification (all modules) type (models, interfaces)
+### Fixed
+- Workflow status fixed
+- Zones added in BaseModel
+- update input are Partial
 ## 2.5.4 - 2023-08-12
 
 ## 2.5.3 - 2023-08-12